Add files via upload
adungeon.c uses command line args to create a basic dungeon structure and displays it using ascii characters dumping it all into the terminal after generated. 
usage: ./adungeon [max_blocks] [place_probability] [seed]

adungeon_interactive.c prompts the user for each of the args and explains what each is used for during run time. It also prompts the user for the name of the text file the result is dumped into. blank space is displayed as # but the rest of the elements are displayed as ints 1-4.
usage: ./adungeon_interactive